make the pong game code cleaner
