+++ backend/src/notification/notification.module.ts
@@ -2,7 +2,7 @@ import { Global, Module } from '@nestjs/common';
 import { NotificationGateway } from './notification.gateway';
 import { NotificationController } from './notification.controller';
 import NotificationService from './notification.service';
-
+@Global()
 @Module({
   providers: [NotificationService, NotificationGateway],
   exports: [NotificationService],
