+++ backend/src/pong/classes/index.ts
@@ -1,0 +1,5 @@
+export * from './ai-player';
+export * from './ball';
+export * from './player';
+export * from './game';
+export * from './canvas';
