+++ backend/src/pong/interfaces/index.ts
@@ -1,0 +1,9 @@
+import { Socket } from 'socket.io';
+
+export interface Invitation {
+  inviterId: number;
+  inviterSocket?: Socket;
+  invitedFriendId: number;
+  invitedFriendSocket?: Socket;
+  timestamp: number;
+}
