+++ backend/src/pong/services/gameprovider.service.ts
@@ -1,0 +1,132 @@
+import { Canvas, Game, Player } from '../classes';
+
+class GameProvider {
+  public id: number;
+  public game: Game;
+  public gameStarted: boolean = false;
+  public ballChangedDirection = false;
+  public ballDirection: number = -1;
+  constructor() {
+    // this.game = null;
+    // const id = setTimeout(() => {
+    this.gameStarted = true;
+    //   clearTimeout(id);
+    // }, 5000);
+  }
+
+  init(playerA: Player, playerB: Player) {
+    this.game = new Game(playerA, playerB);
+    playerB.x = playerB.canvas.width - playerB.width;
+    return this.game;
+  }
+
+  paddleCol(player: Player) {
+    const ball = this.game.ball;
+    const paddleRight = player.x + player.width;
+    const paddleLeft = player.x;
+    const paddleTop = player.y;
+    const paddleBottom = player.y + player.height;
+    return (
+      ball.x - ball.radius < paddleRight &&
+      ball.x + ball.radius > paddleLeft &&
+      ball.y - ball.radius < paddleBottom &&
+      ball.y + ball.radius > paddleTop
+    );
+  }
+
+  update(info: { userId: number; playerCanvas: Canvas }) {
+    if (!this.gameStarted) return;
+    const { userId, playerCanvas } = info;
+    const playerA = this.game.playerA;
+    const playerB = this.game.playerB;
+    const ball = this.game.ball;
+    const { width, height } = playerCanvas;
+
+    if (!this.game || !playerB) return;
+
+    const player = userId === playerA.id ? playerA : playerB;
+
+    if (playerA.keyState['ArrowUp'] && playerA.y > 5) {
+      playerA.y -= 5;
+    } else if (
+      playerA.keyState['ArrowDown'] &&
+      playerA.y < height - playerA.height - 5
+    ) {
+      playerA.y += 5;
+    }
+
+    if (playerB.keyState['ArrowUp'] && playerB.y > 5) {
+      playerB.y -= 5;
+    } else if (
+      playerB.keyState['ArrowDown'] &&
+      playerB.y < height - playerB.height - 5
+    ) {
+      playerB.y += 5;
+    }
+
+    ball.x += (ball.velocity.x * ball.speed) / 10;
+    ball.y += (ball.velocity.y * ball.speed) / 10;
+
+    // Check if ball hits top or bottom wall
+    if (ball.y + ball.radius > height - 10 || ball.y - ball.radius < 10) {
+      ball.velocity.y = -ball.velocity.y;
+    }
+
+    // Check if ball hits player A or B
+    if (
+      (this.paddleCol(this.game.playerA) ||
+        this.paddleCol(this.game.playerB)) &&
+      !this.ballChangedDirection
+    ) {
+      this.ballChangedDirection = true;
+      ball.velocity.x = 2 * -Math.sign(ball.velocity.x);
+      ball.velocity.y = Math.random() * 4 - 2;
+      ball.velocity.x +=
+        Math.sign(ball.velocity.x) * (2 - Math.abs(ball.velocity.y));
+      ball.speed++;
+      // console.log(ball.velocity);
+      const timeoutId = setTimeout(() => {
+        this.ballChangedDirection = false;
+        clearTimeout(timeoutId);
+      }, 1000);
+    }
+
+    // Check if ball goes out of bounds (player A scores)
+    if (ball.x > width) {
+      playerA.score++;
+      ball.reset(playerA.canvas);
+      ball.velocity.x = Math.abs(ball.velocity.x) * this.ballDirection;
+      this.ballDirection = -this.ballDirection;
+    }
+
+    // Check if ball goes out of bounds (player B scores)
+    if (ball.x < 0) {
+      playerB.score++;
+      ball.reset(playerB.canvas);
+      ball.velocity.x = Math.abs(ball.velocity.x) * this.ballDirection;
+      this.ballDirection = -this.ballDirection;
+    }
+
+    return this.game;
+  }
+
+  keyPressed(info: { key: string; userId: number }) {
+    const player =
+      info.userId === this.game.playerA.id
+        ? this.game.playerA
+        : this.game.playerB;
+    player.keyState[info.key] = true;
+    return this.game;
+  }
+
+  keyReleased(info: { key: string; userId: number }) {
+    const player =
+      info.userId === this.game.playerA.id
+        ? this.game.playerA
+        : this.game.playerB;
+    player.keyState[info.key] = false;
+    return this.game;
+  }
+}
+
+export { GameProvider };
